Refactor algorithms to support Float32 data type
- Updated ADMMSLIM, EALS, EASE, FM, FTRL, GloVe, ItemKNN, SLIM, SoftImpute algorithms to accept a dtype parameter, defaulting to Float32.
- Modified matrix operations in recommend functions to ensure compatibility with the specified dtype.
- Adjusted tests to validate the output types and ensure consistency with Float32.
- Removed outdated benchmark output file and updated benchmark configurations for small and medium scales.

+++ src/algorithms/eals.jl
@@ -86,7 +86,7 @@ function EALS(;
     convergence_tol::Float64 = 0.005,
     popularity_exponent::Float64 = 0.5,
     verbose::Bool = true,
-    dtype::Type{<:AbstractFloat} = Float64,
+    dtype::Type{<:AbstractFloat} = Float32,
 )
     rank >= 1 || throw(ArgumentError("rank must be ≥ 1, got $rank"))
     λ >= 0.0 || throw(ArgumentError("λ must be non-negative, got $λ"))
